ProofOfConcept bugfixes and weights
Fixed a few issues that would occur if both the true and false votes for
an element were 0, and some bugs in the random mode.

Random mode now enabled by default in the text file, with larger numbers
as well just for proof of concept

Weights are now calculated but this is something I'm still very much
messing around with.  The current weighting code in here isn't even the
same as my scratch work or excel spreadsheet, although I did still use
those to confirm code functionality.