Hesitant Optimism
As far as I can tell, everything that it's programmed to do gets done
correctly.  It doesn't quite have the functionality completed but the
remaining stuff, at least as far as what I had planned for this little
test program, should be pretty trivial to add.

The comments in this are probably atrocious because I'm working under
the assumption that (practically) nobody else will ever be reading this.
I'll definitely clean that all up when I wrap up the programming though.
There's also a bit more efficiency improvement work to be done.