List Completion and code cleanup
The program now goes through (efficiently) the voted rules for the list
as many times as is needed to achieve completion.  Also, I did my best
to clean up many of the comments and such.  I wasn't trying to be
extremely scrutinous so there very likely are still some issues with
that.