Bug free (?) ordering
The program doesn't seem to be giving me any errors even after running
it hundreds of times with randomly generated data so I'm assuming it
works every time (at least for lists of length 5 and up, there could
very well be a minimum list size and I'll be sure to check that)

The code is cleaned up pretty well and is decently efficient to my
knowledge.  Next I'm going to start adding more functionality for manual
file input and working on analyzing the generated list to figure out how
"good" of a solution it is.