Commit to save the code in its current state
The program is very buggy right now, it doesn't even work for length 5
lists anymore actually. The further I got into my method of finding the
best way to order the items once condition weights are calculated, the
more evident it became that this problem was much more complex than I
had initially thought.  This pushed me to consider other routes of
computation, and I am preparing to remove much of the code I've written
thus far and replace it with more straightforward functions (which work
better anyway).  I'm committing this because I don't want to lose any of
the work, just in case I decide to go back to this method.