Begin overhaul of sorting process
As stated in the last commit, a ton of code is getting removed here.  I
wrote a new Comparison class which will make a lot of the sorting work
easier to follow and just more intuitive to write in general.  None of
the actual new algorithm is implemented yet, I'm just setting the stage
for that work to be done.