Implement new sorting algorithm
The algorithm I worked on yesterday has been implemented into the code
and doesn't appear to give any errors.  Not that I know that it's
bug-free as far as ordering, but from a glance-over of many sets of
results it appears pretty functional.  From here I can work on the
weighting algorithm and some minor sort tweaks to try to give the
highest number of people satisfied as is possible.

Note: The code is very output-heavy right now