Implement statistics
I added some basic analysis just to be able to tweak the algorithm and
get an idea of how "satisfied" the final list can be. The major
limitation to being able to tweak the algorithm at this point is that
the randomly generated data doesn't reflect any sort of "individual
opinions" in any way.