Implement item ordering
This commit includes the work I've done so far to allow for the items in
an ItemList to have their order_index values calculated.  The code in
Concept.py is heavily based off of earlier work done in
ProofOfConcept.py

I don't expect this code to be bug free but I am done for the night and
probably won't work on this tomorrow so I wanted to leave off somewhere
acceptable.  On a less interesting note, I need to figure out how to
properly import Concept.py into views.py